scripts/python/apiview-syncdb/requirements.txt
azure-cosmos